Add quadrotor SOC example
