+++ examples/Project.toml
@@ -1,4 +1,5 @@
 [deps]
+COSMO = "1e616198-aa4e-51ec-90a2-23f7fbd31d8d"
 COSMOAccelerators = "bbd8fffe-5ad0-4d78-a55e-85575421b4ac"
 Colors = "5ae59095-9a9b-59fe-a467-6f913c188581"
 CoordinateTransformations = "150eb455-5306-5404-9cee-2592286d6298"
